Rename `endpoints` modules to `views`
